+++ packages/rum-core/src/domain/configuration.spec.ts
@@ -1,5 +1,6 @@
 import { DefaultPrivacyLevel, display } from '@datadog/browser-core'
-import { validateAndBuildRumConfiguration } from './configuration'
+import type { RumInitConfiguration } from './configuration'
+import { serializeRumConfiguration, validateAndBuildRumConfiguration } from './configuration'
 
 const DEFAULT_INIT_CONFIGURATION = { clientToken: 'xxx', applicationId: 'xxx' }
 
@@ -166,37 +167,36 @@ describe('validateAndBuildRumConfiguration', () => {
   })
 
   describe('allowedTracingOrigins', () => {
-    it('defaults to an empty array', () => {
-      expect(validateAndBuildRumConfiguration(DEFAULT_INIT_CONFIGURATION)!.allowedTracingOrigins).toEqual([])
-    })
-
     it('is set to provided value', () => {
       expect(
         validateAndBuildRumConfiguration({
           ...DEFAULT_INIT_CONFIGURATION,
           allowedTracingOrigins: ['foo'],
           service: 'bar',
-        })!.allowedTracingOrigins
-      ).toEqual(['foo'])
+        })!.allowedTracingUrls
+      ).toEqual([{ match: 'foo', propagatorTypes: ['datadog'] }])
     })
 
     it('accepts functions', () => {
-      const customOriginFunction = (origin: string): boolean => origin === 'foo'
+      const originMatchSpy = jasmine.createSpy<(origin: string) => boolean>()
 
-      expect(
-        validateAndBuildRumConfiguration({
-          ...DEFAULT_INIT_CONFIGURATION,
-          allowedTracingOrigins: [customOriginFunction],
-          service: 'bar',
-        })!.allowedTracingOrigins
-      ).toEqual([customOriginFunction])
+      const tracingUrlOptionMatch = validateAndBuildRumConfiguration({
+        ...DEFAULT_INIT_CONFIGURATION,
+        allowedTracingOrigins: [originMatchSpy],
+        service: 'bar',
+      })!.allowedTracingUrls[0].match as (url: string) => boolean
+
+      expect(typeof tracingUrlOptionMatch).toBe('function')
+      // Replicating behavior from allowedTracingOrigins, new function will treat the origin part of the URL
+      tracingUrlOptionMatch('https://my.origin.com/api')
+      expect(originMatchSpy).toHaveBeenCalledWith('https://my.origin.com')
     })
 
     it('does not validate the configuration if a value is provided and service is undefined', () => {
       expect(
         validateAndBuildRumConfiguration({ ...DEFAULT_INIT_CONFIGURATION, allowedTracingOrigins: ['foo'] })
       ).toBeUndefined()
-      expect(displayErrorSpy).toHaveBeenCalledOnceWith('Service need to be configured when tracing is enabled')
+      expect(displayErrorSpy).toHaveBeenCalledOnceWith('Service needs to be configured when tracing is enabled')
     })
 
     it('does not validate the configuration if an incorrect value is provided', () => {
@@ -204,6 +204,85 @@ describe('validateAndBuildRumConfiguration', () => {
         validateAndBuildRumConfiguration({ ...DEFAULT_INIT_CONFIGURATION, allowedTracingOrigins: 'foo' as any })
       ).toBeUndefined()
       expect(displayErrorSpy).toHaveBeenCalledOnceWith('Allowed Tracing Origins should be an array')
+    })
+  })
+
+  describe('allowedTracingUrls', () => {
+    it('defaults to an empty array', () => {
+      expect(validateAndBuildRumConfiguration(DEFAULT_INIT_CONFIGURATION)!.allowedTracingUrls).toEqual([])
+    })
+
+    it('is set to provided value', () => {
+      expect(
+        validateAndBuildRumConfiguration({
+          ...DEFAULT_INIT_CONFIGURATION,
+          allowedTracingUrls: ['foo'],
+          service: 'bar',
+        })!.allowedTracingUrls
+      ).toEqual([{ match: 'foo', propagatorTypes: ['datadog'] }])
+    })
+
+    it('accepts functions', () => {
+      const customOriginFunction = (url: string): boolean => url === 'https://my.origin.com'
+
+      expect(
+        validateAndBuildRumConfiguration({
+          ...DEFAULT_INIT_CONFIGURATION,
+          allowedTracingUrls: [customOriginFunction],
+          service: 'bar',
+        })!.allowedTracingUrls
+      ).toEqual([{ match: customOriginFunction, propagatorTypes: ['datadog'] }])
+    })
+
+    it('accepts RegExp', () => {
+      expect(
+        validateAndBuildRumConfiguration({
+          ...DEFAULT_INIT_CONFIGURATION,
+          allowedTracingUrls: [/az/i],
+          service: 'bar',
+        })!.allowedTracingUrls
+      ).toEqual([{ match: /az/i, propagatorTypes: ['datadog'] }])
+    })
+
+    it('keeps headers', () => {
+      expect(
+        validateAndBuildRumConfiguration({
+          ...DEFAULT_INIT_CONFIGURATION,
+          allowedTracingUrls: [{ match: 'simple', propagatorTypes: ['b3multi', 'tracecontext'] }],
+          service: 'bar',
+        })!.allowedTracingUrls
+      ).toEqual([{ match: 'simple', propagatorTypes: ['b3multi', 'tracecontext'] }])
+    })
+
+    it('should filter out unexpected parameter types', () => {
+      expect(
+        validateAndBuildRumConfiguration({
+          ...DEFAULT_INIT_CONFIGURATION,
+          service: 'bar',
+          allowedTracingUrls: [
+            42 as any,
+            undefined,
+            { match: 42 as any, propagatorTypes: ['datadog'] },
+            { match: 'toto' },
+          ],
+        })!.allowedTracingUrls
+      ).toEqual([])
+
+      expect(displayWarnSpy).toHaveBeenCalledTimes(4)
+    })
+
+    it('does not validate the configuration if a value is provided and service is undefined', () => {
+      expect(
+        validateAndBuildRumConfiguration({ ...DEFAULT_INIT_CONFIGURATION, allowedTracingUrls: ['foo'] })
+      ).toBeUndefined()
+      expect(displayErrorSpy).toHaveBeenCalledOnceWith('Service needs to be configured when tracing is enabled')
+    })
+
+    it('does not validate the configuration if an incorrect value is provided', () => {
+      expect(
+        validateAndBuildRumConfiguration({ ...DEFAULT_INIT_CONFIGURATION, allowedTracingUrls: 'foo' as any })
+      ).toBeUndefined()
+      expect(displayErrorSpy).toHaveBeenCalledOnceWith('Allowed Tracing URLs should be an array')
     })
   })
 
@@ -380,6 +459,38 @@ describe('validateAndBuildRumConfiguration', () => {
       expect(
         validateAndBuildRumConfiguration({ ...DEFAULT_INIT_CONFIGURATION, trackLongTasks: false })!.trackLongTasks
       ).toBeFalse()
+    })
+  })
+
+  describe('serializeRumConfiguration', () => {
+    describe('selected tracing propagators serialization', () => {
+      it('should not return any propagator type', () => {
+        expect(serializeRumConfiguration(DEFAULT_INIT_CONFIGURATION).selected_tracing_propagators).toEqual([])
+      })
+
+      it('should return Datadog propagator type', () => {
+        const simpleTracingConfig: RumInitConfiguration = {
+          ...DEFAULT_INIT_CONFIGURATION,
+          allowedTracingUrls: ['foo'],
+        }
+        expect(serializeRumConfiguration(simpleTracingConfig).selected_tracing_propagators).toEqual(['datadog'])
+      })
+
+      it('should return all propagator types', () => {
+        const complexTracingConfig: RumInitConfiguration = {
+          ...DEFAULT_INIT_CONFIGURATION,
+          allowedTracingUrls: [
+            'foo',
+            { match: 'first', propagatorTypes: ['datadog'] },
+            { match: 'test', propagatorTypes: ['tracecontext'] },
+            { match: 'other', propagatorTypes: ['b3'] },
+            { match: 'final', propagatorTypes: ['b3multi'] },
+          ],
+        }
+        expect(serializeRumConfiguration(complexTracingConfig).selected_tracing_propagators).toEqual(
+          jasmine.arrayWithExactContents(['datadog', 'b3', 'b3multi', 'tracecontext'])
+        )
+      })
     })
   })
 })