+++ packages/rum-core/src/domain/configuration.ts
@@ -1,5 +1,8 @@
 import type { Configuration, InitConfiguration, MatchOption, RawTelemetryConfiguration } from '@datadog/browser-core'
 import {
+  arrayFrom,
+  getOrigin,
+  isMatchOption,
   serializeConfiguration,
   assign,
   DefaultPrivacyLevel,
@@ -10,6 +13,8 @@ import {
 } from '@datadog/browser-core'
 import type { RumEventDomainContext } from '../domainContext.types'
 import type { RumEvent } from '../rumEvent.types'
+import { isTracingOption } from './tracing/tracer'
+import type { PropagatorType, TracingOption } from './tracing/tracer.types'
 
 export interface RumInitConfiguration extends InitConfiguration {
   // global options
@@ -22,7 +27,11 @@ export interface RumInitConfiguration extends InitConfiguration {
   excludedActivityUrls?: MatchOption[] | undefined
 
   // tracing options
+  /**
+   * @deprecated use allowedTracingUrls instead
+   */
   allowedTracingOrigins?: MatchOption[] | undefined
+  allowedTracingUrls?: Array<MatchOption | TracingOption> | undefined
   tracingSampleRate?: number | undefined
 
   // replay options
@@ -50,8 +59,8 @@ export type HybridInitConfiguration = Omit<RumInitConfiguration, 'applicationId'
 export interface RumConfiguration extends Configuration {
   // Built from init configuration
   actionNameAttribute: string | undefined
-  allowedTracingOrigins: MatchOption[]
   tracingSampleRate: number | undefined
+  allowedTracingUrls: TracingOption[]
   excludedActivityUrls: MatchOption[]
   applicationId: string
   defaultPrivacyLevel: DefaultPrivacyLevel
@@ -98,19 +107,13 @@ export function validateAndBuildRumConfiguration(
     return
   }
 
-  if (initConfiguration.allowedTracingOrigins !== undefined) {
-    if (!Array.isArray(initConfiguration.allowedTracingOrigins)) {
-      display.error('Allowed Tracing Origins should be an array')
-      return
-    }
-    if (initConfiguration.allowedTracingOrigins.length !== 0 && initConfiguration.service === undefined) {
-      display.error('Service need to be configured when tracing is enabled')
-      return
-    }
-  }
-
   if (initConfiguration.excludedActivityUrls !== undefined && !Array.isArray(initConfiguration.excludedActivityUrls)) {
     display.error('Excluded Activity Urls should be an array')
+    return
+  }
+
+  const allowedTracingUrls = validateAndBuildTracingOptions(initConfiguration)
+  if (!allowedTracingUrls) {
     return
   }
 
@@ -128,8 +131,8 @@ export function validateAndBuildRumConfiguration(
       actionNameAttribute: initConfiguration.actionNameAttribute,
       sessionReplaySampleRate: initConfiguration.sessionReplaySampleRate ?? premiumSampleRate ?? 100,
       oldPlansBehavior: initConfiguration.sessionReplaySampleRate === undefined,
-      allowedTracingOrigins: initConfiguration.allowedTracingOrigins ?? [],
       tracingSampleRate: initConfiguration.tracingSampleRate,
+      allowedTracingUrls,
       excludedActivityUrls: initConfiguration.excludedActivityUrls ?? [],
       trackInteractions: !!initConfiguration.trackInteractions || trackFrustrations,
       trackFrustrations,
@@ -144,6 +147,114 @@ export function validateAndBuildRumConfiguration(
   )
 }
 
+/**
+ * Handles allowedTracingUrls and processes legacy allowedTracingOrigins
+ */
+function validateAndBuildTracingOptions(initConfiguration: RumInitConfiguration): TracingOption[] | undefined {
+  // Advise about parameters precedence.
+  if (initConfiguration.allowedTracingUrls !== undefined && initConfiguration.allowedTracingOrigins !== undefined) {
+    display.warn(
+      'Both allowedTracingUrls and allowedTracingOrigins (deprecated) have been defined. The parameter allowedTracingUrls will override allowedTracingOrigins.'
+    )
+  }
+  // Handle allowedTracingUrls first
+  if (initConfiguration.allowedTracingUrls !== undefined) {
+    if (!Array.isArray(initConfiguration.allowedTracingUrls)) {
+      display.error('Allowed Tracing URLs should be an array')
+      return
+    }
+    if (initConfiguration.allowedTracingUrls.length !== 0 && initConfiguration.service === undefined) {
+      display.error('Service needs to be configured when tracing is enabled')
+      return
+    }
+    // Convert from (MatchOption | TracingOption) to TracingOption, remove unknown properties
+    const tracingOptions: TracingOption[] = []
+    initConfiguration.allowedTracingUrls.forEach((option) => {
+      if (isMatchOption(option)) {
+        tracingOptions.push({ match: option, propagatorTypes: ['datadog'] })
+      } else if (isTracingOption(option)) {
+        tracingOptions.push(option)
+      } else {
+        display.warn(
+          'Allowed Tracing Urls parameters should be a string, RegExp, function, or an object. Ignoring parameter',
+          option
+        )
+      }
+    })
+
+    return tracingOptions
+  }
+
+  // Handle conversion of allowedTracingOrigins to allowedTracingUrls
+  if (initConfiguration.allowedTracingOrigins !== undefined) {
+    if (!Array.isArray(initConfiguration.allowedTracingOrigins)) {
+      display.error('Allowed Tracing Origins should be an array')
+      return
+    }
+    if (initConfiguration.allowedTracingOrigins.length !== 0 && initConfiguration.service === undefined) {
+      display.error('Service needs to be configured when tracing is enabled')
+      return
+    }
+
+    const tracingOptions: TracingOption[] = []
+    initConfiguration.allowedTracingOrigins.forEach((legacyMatchOption) => {
+      const tracingOption = convertLegacyMatchOptionToTracingOption(legacyMatchOption)
+      if (tracingOption) {
+        tracingOptions.push(tracingOption)
+      }
+    })
+    return tracingOptions
+  }
+
+  return []
+}
+
+/**
+ * Converts parameters from the deprecated allowedTracingOrigins
+ * to allowedTracingUrls. Handles the change from origin to full URLs.
+ */
+function convertLegacyMatchOptionToTracingOption(item: MatchOption): TracingOption | undefined {
+  let match: MatchOption | undefined
+  if (typeof item === 'string') {
+    match = item
+  } else if (item instanceof RegExp) {
+    match = (url) => item.test(getOrigin(url))
+  } else if (typeof item === 'function') {
+    match = (url) => item(getOrigin(url))
+  }
+
+  if (match === undefined) {
+    display.warn('Allowed Tracing Origins parameters should be a string, RegExp or function. Ignoring parameter', item)
+    return undefined
+  }
+
+  return { match, propagatorTypes: ['datadog'] }
+}
+
+/**
+ * Combines the selected tracing propagators from the different options in allowedTracingUrls,
+ * and assumes 'datadog' has been selected when using allowedTracingOrigins
+ */
+function getSelectedTracingPropagators(configuration: RumInitConfiguration): PropagatorType[] {
+  const usedTracingPropagators = new Set<PropagatorType>()
+
+  if (Array.isArray(configuration.allowedTracingUrls) && configuration.allowedTracingUrls.length > 0) {
+    configuration.allowedTracingUrls.forEach((option) => {
+      if (isMatchOption(option)) {
+        usedTracingPropagators.add('datadog')
+      } else {
+        option.propagatorTypes.forEach((propagatorType) => usedTracingPropagators.add(propagatorType))
+      }
+    })
+  }
+
+  if (Array.isArray(configuration.allowedTracingOrigins) && configuration.allowedTracingOrigins.length > 0) {
+    usedTracingPropagators.add('datadog')
+  }
+
+  return arrayFrom(usedTracingPropagators)
+}
+
 export function serializeRumConfiguration(configuration: RumInitConfiguration): RawTelemetryConfiguration {
   const baseSerializedConfiguration = serializeConfiguration(configuration)
 
@@ -156,6 +267,9 @@ export function serializeRumConfiguration(configuration: RumInitConfiguration): 
       action_name_attribute: configuration.actionNameAttribute,
       use_allowed_tracing_origins:
         Array.isArray(configuration.allowedTracingOrigins) && configuration.allowedTracingOrigins.length > 0,
+      use_allowed_tracing_urls:
+        Array.isArray(configuration.allowedTracingUrls) && configuration.allowedTracingUrls.length > 0,
+      selected_tracing_propagators: getSelectedTracingPropagators(configuration),
       default_privacy_level: configuration.defaultPrivacyLevel,
       use_excluded_activity_urls:
         Array.isArray(configuration.allowedTracingOrigins) && configuration.allowedTracingOrigins.length > 0,