+++ packages/rum-core/src/domain/rumEventsCollection/resource/resourceCollection.spec.ts
@@ -96,9 +96,6 @@ describe('resourceCollection', () => {
       type: RumEventType.RESOURCE,
       _dd: {
         discarded: false,
-        resolveDuration: undefined,
-        durationDiff: undefined,
-        durationPercentageDiff: undefined,
       },
     })
     expect(rawRumEvents[0].domainContext).toEqual({
@@ -122,7 +119,6 @@ describe('resourceCollection', () => {
       LifeCycleEventType.REQUEST_COMPLETED,
       createCompletedRequest({
         duration: 500 as Duration,
-        resolveDuration: 50 as Duration,
         method: 'GET',
         startClocks: relativeToClocks(100 as RelativeTime),
         status: 200,
@@ -150,7 +146,6 @@ describe('resourceCollection', () => {
       LifeCycleEventType.REQUEST_COMPLETED,
       createCompletedRequest({
         duration: 100 as Duration,
-        resolveDuration: 50 as Duration,
         method: 'GET',
         startClocks: { relative: 1234 as RelativeTime, timeStamp: 123456789 as TimeStamp },
         status: 200,
@@ -176,9 +171,6 @@ describe('resourceCollection', () => {
       type: RumEventType.RESOURCE,
       _dd: {
         discarded: false,
-        resolveDuration: (50 * 1e6) as ServerDuration,
-        durationDiff: (50 * 1e6) as ServerDuration,
-        durationPercentageDiff: 50,
       },
     })
     expect(rawRumEvents[0].domainContext).toEqual({